requirements.txt

stravalib==1.2.0
pyyaml
boto3
psycopg2
spotipy
requests_oauthlib
schedule
argparse
python-crontab
pandas
requests

